articles/pyproject.toml
[tool.poetry]
name = "articles"
version = "0.1.0"
description = ""
authors = ["aliraad79 <a.ahmadi.k.79@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
django-redis = "^5.4.0"
gunicorn = "^22.0.0"
django-environ = "^0.11.2"
requests = "^2.32.3"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
